+++ src/main/java/com/asu/alliancebank/factory/impl/TransferFundsFactory.java
@@ -28,7 +28,7 @@ public class TransferFundsFactory implements ITransferFundsFactory{
 				
 		return new TransferFunds(transferFundsBackingBean.getFromAccountId(),			
 				transferFundsBackingBean.getToAccountId(),
-				Long.parseLong(transferFundsBackingBean.getAmount()));				
+				 Double.parseDouble(transferFundsBackingBean.getAmount()));				
 	}
 
 	@Override
