+++ src/main/java/com/asu/alliancebank/service/transaction/impl/TransferFundsManager.java
@@ -15,13 +15,10 @@ import org.springframework.stereotype.Service;
 import com.asu.alliancebank.db.transaction.ITransferFundsDBManager;
 import com.asu.alliancebank.domain.ITransactionTransferFund;
 import com.asu.alliancebank.domain.impl.TransferFunds;
-import com.asu.alliancebank.domain.impl.User;
 import com.asu.alliancebank.security.otp.impl.OTPManager;
 import com.asu.alliancebank.service.account.IAccountManager;
-import com.asu.alliancebank.service.role.IRoleManager;
 import com.asu.alliancebank.service.transaction.ITransferFundsManager;
 import com.asu.alliancebank.service.user.IUserManager;
-import com.asu.alliancebank.service.userservice.AllianceBankGrantedAuthority;
 
 @Service
 public class TransferFundsManager implements ITransferFundsManager{
@@ -71,7 +68,7 @@ public class TransferFundsManager implements ITransferFundsManager{
 	@Override
 	public boolean isValid(String loggedInUser, String amount)
 			throws SQLException {
-		Long balance = null;
+		Double balance = null;
 		try {
 			balance = accountManager.getAccountBalance(loggedInUser);
 		} catch (SQLException e1) {
@@ -79,7 +76,7 @@ public class TransferFundsManager implements ITransferFundsManager{
 		}
 
 		try {
-			Long amountLong = Long.parseLong(amount);
+			double amountLong =  Double.parseDouble(amount);
 			if (balance - amountLong > 0) {
 				return true;
 			}
